Del with backup

+++ lr01-batch-user-reg/generate_users_list.sh
@@ -4,9 +4,17 @@ dir=`dirname $0`
 
 result_file="$dir/users_list.txt"
 
-count_users=500
+if [ $# -ne 0 ];
+then
+count_users=$1
+else
+count_users=100
+fi
 
+if [ -f $result_file ];
+then
 rm $result_file
+fi
 
 for i in $(seq -w 1 $count_users);
 do
